+++ tests/src/main/java/jme3utilities/debug/test/TestPas.java
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 2014-2017, Stephen Gold
+ Copyright (c) 2014-2018, Stephen Gold
  All rights reserved.
 
  Redistribution and use in source and binary forms, with or without
@@ -27,6 +27,7 @@
 package jme3utilities.debug.test;
 
 import com.jme3.app.SimpleApplication;
+import com.jme3.app.StatsAppState;
 import com.jme3.audio.openal.ALAudioRenderer;
 import com.jme3.system.AppSettings;
 import java.util.logging.Level;
@@ -91,5 +92,9 @@ public class TestPas extends SimpleApplication {
     public void simpleInitApp() {
         PerformanceAppState pas = new PerformanceAppState();
         stateManager.attach(pas);
+        /*
+         * Detach any JME stats app state(s).
+         */
+        Misc.detachAll(stateManager, StatsAppState.class);
     }
 }